Update redtexts to add automated ATOM feed generation! Details below.
This involved two changes to every text file. First, each text file gets a new metadata line, Atomdate. This contains the file's posting date in a format suitable to be inserted straight into an atom feed. And a line was added to the HTML header metadata field so that each html page has a line it its header declaring that there is an atom feed for the site.

The other pieces of this were: to update the index_header.txt and article_header.txt files to include an RSS link; to add a first draft of a script that could look at each file and generate an atom feed from them; to update the template file; and to update the modified field of all files.