End description metadata fields with periods.
A bit of history to explain why they were absent: in the first version of redtexts, the HTML files were generated using a template, in which the "description" field was a variable that got inserted into a line that already included a period. It looked something like this: "%Description%. Added on %Added.%..." That meant that including a period in the description metadata would've resulted in an HTML file with two periods between the description sentence(s) and the "Added on..." sentence.