Fix many errors from the original scan 20 years ago. Use comment metadata to point a scan on archive.org that can be used for further proofreading.
As an aside, I will point out that particular errors in the spunk.org version (e.g., Boat instead of float) crop up in many online versions. Pointing readers to those is not as useful as pointing to a PDF scan of the original text. Also, the original printed text used a lot of italics. We haven't added those so far.