Add Philoren - Money Must Go!
This isn't the greatest text but it did not exist online as a PDF or plain text, as far as I can tell. I uploaded it to libcom and archive.org a couple of days ago, but this red texts version should be regarded as the canonical version. It can be improved and tracked."